pyproject.toml
[project]
name = "video-search-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.10",
    "aiosignal==1.3.1",
    "altair==5.4.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appnope==0.1.4",
    "asgiref==3.8.1",
    "asttokens==3.0.0",
    "attrs==24.2.0",
    "backoff==2.2.1",
    "bcrypt==4.2.0",
    "beanie==1.29.0",
    "blinker==1.8.2",
    "boto3==1.36.21",
    "botocore==1.36.21",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.7.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.5.20",
    "click==8.1.7",
    "coloredlogs==15.0.1",
    "comm==0.2.2",
    "cramjam==2.9.1",
    "cryptography==44.0.1",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.13",
    "decorator==5.2.1",
    "deprecated==1.2.14",
    "deptry==0.23.0",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "docstring-parser==0.16",
    "durationpy==0.9",
    "email-validator==2.2.0",
    "executing==2.2.0",
    "fastapi-cli==0.0.6",
    "fastapi==0.115.4",
    "filelock==3.16.1",
    "filetype==1.2.0",
    "flashrank==0.2.9",
    "flatbuffers==24.3.25",
    "frozenlist==1.5.0",
    "fsspec==2024.10.0",
    "gitdb==4.0.11",
    "gitpython==3.1.43",
    "google-ai-generativelanguage==0.6.10",
    "google-api-core==2.22.0",
    "google-api-python-client==2.151.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-cloud-aiplatform==1.82.0",
    "google-cloud-bigquery==3.30.0",
    "google-cloud-core==2.4.2",
    "google-cloud-resource-manager==1.14.1",
    "google-cloud-storage==2.19.0",
    "google-crc32c==1.6.0",
    "google-genai==1.25.0",
    "google-generativeai==0.8.3",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.65.0",
    "groq==0.11.0",
    "grpc-google-iam-v1==0.14.1",
    "grpcio==1.71.0",
    "grpcio-status==1.67.1",
    "grpcio-tools==1.71.0",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.26.2",
    "humanfriendly==10.0",
    "hyperframe==6.1.0",
    "idna==3.10",
    "importlib-metadata==8.5.0",
    "importlib-resources==6.4.5",
    "ipykernel==6.29.5",
    "ipython==9.0.2",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.4",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "kubernetes==31.0.0",
    "langchain==0.3.7",
    "langchain-chroma==0.1.4",
    "langchain-community==0.3.5",
    "langchain-core==0.3.41",
    "langchain-google-genai==2.0.9",
    "langchain-google-vertexai==2.0.14",
    "langchain-groq==0.2.1",
    "langchain-huggingface==0.1.2",
    "langchain-text-splitters==0.3.2",
    "langsmith==0.1.141",
    "lazy-model==0.2.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.23.1",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mmh3==5.0.1",
    "monotonic==1.6",
    "motor==3.7.0",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "narwhals==1.13.2",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "onnxruntime==1.20.0",
    "opentelemetry-api==1.28.0",
    "opentelemetry-exporter-otlp-proto-common==1.28.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.28.0",
    "opentelemetry-instrumentation==0.49b0",
    "opentelemetry-instrumentation-asgi==0.49b0",
    "opentelemetry-instrumentation-fastapi==0.49b0",
    "opentelemetry-proto==1.28.0",
    "opentelemetry-sdk==1.28.0",
    "opentelemetry-semantic-conventions==0.49b0",
    "opentelemetry-util-http==0.49b0",
    "orjson==3.10.11",
    "overrides==7.7.0",
    "packaging==24.1",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.0.0",
    "pip==24.2",
    "platformdirs==4.3.7",
    "portalocker==2.10.1",
    "posthog==3.7.0",
    "prompt-toolkit==3.0.50",
    "propcache==0.2.0",
    "proto-plus==1.25.0",
    "protobuf==5.28.3",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==18.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.6.1",
    "pydeck==0.9.1",
    "pygments==2.18.0",
    "pykerberos==1.2.4",
    "pymongo==4.11.1",
    "pymongo-auth-aws==1.3.0",
    "pymongocrypt==1.12.2",
    "pyopenssl==25.0.0",
    "pyparsing==3.2.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "python-dotenv==1.0.1",
    "python-magic==0.4.27",
    "python-multipart==0.0.19",
    "python-snappy==0.7.3",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "pyzmq==26.3.0",
    "rank-bm25==0.2.2",
    "referencing==0.35.1",
    "regex==2024.11.6",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "requirements-parser==0.13.0",
    "rich==13.9.4",
    "rich-toolkit==0.11.3",
    "rpds-py==0.21.0",
    "rsa==4.9.1",
    "s3transfer==0.11.2",
    "safetensors==0.4.5",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "sentence-transformers==3.2.1",
    "sentry-sdk==2.21.0",
    "service-identity==24.2.0",
    "setuptools==74.1.2",
    "shapely==2.0.7",
    "shellingham==1.5.4",
    "six==1.16.0",
    "smmap==5.0.1",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.35",
    "stack-data==0.6.3",
    "starlette==0.41.2",
    "streamlit==1.40.0",
    "sympy==1.13.1",
    "tenacity==8.5.0",
    "threadpoolctl==3.5.0",
    "tiktoken==0.8.0",
    "tokenizers==0.20.3",
    "toml==0.10.2",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "tornado==6.4.1",
    "tqdm==4.67.0",
    "traitlets==5.14.3",
    "transformers==4.46.2",
    "typer==0.13.0",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2024.2",
    "uritemplate==4.1.1",
    "urllib3==2.5.0",
    "uvicorn==0.32.0",
    "uvloop==0.21.0",
    "watchfiles==0.24.0",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wrapt==1.16.0",
    "yarl==1.17.1",
    "zipp==3.20.2",
    "zstandard==0.23.0",
    "weaviate-client>=4.16.3",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]
